refactor: 페이지 스타일 분리
