+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
 
-import CardRegistrationPage from './pages/CardRegistrationPage';
-import CardRegistrationCompletePage from './pages/CardRegistrationCompletePage';
+import CardRegistrationPage from './pages/CardRegistrationPage/CardRegistrationPage';
+import CardRegistrationCompletePage from './pages/CardRegistrationCompletePage/CardRegistrationCompletePage';
 
 import theme from './styles/theme';
 import GlobalStyles from './styles/Global.style';
