+++ src/pages/CardRegistrationCompletePage/CardRegistrationCompletePage.style.tsx
@@ -1,7 +1,6 @@
-import { Check } from '../assets';
 import styled from 'styled-components';
 
-const CardRegistrationCompleteLayout = styled.div`
+export const CardRegistrationCompleteLayout = styled.div`
   height: 100vh;
   display: flex;
   flex-direction: column;
@@ -10,7 +9,7 @@ const CardRegistrationCompleteLayout = styled.div`
   align-items: center;
 `;
 
-const CheckBox = styled.div`
+export const CheckBox = styled.div`
   display: flex;
   justify-content: center;
   align-items: center;
@@ -20,14 +19,14 @@ const CheckBox = styled.div`
   background: ${({ theme }) => theme.color.primary.main};
 `;
 
-const CardRegistrationMessage = styled.p`
+export const CardRegistrationMessage = styled.p`
   font-size: ${({ theme }) => theme.fontSize.xl};
   font-weight: 700;
   line-height: 36.2px;
   text-align: center;
 `;
 
-const ConfirmButton = styled.button`
+export const ConfirmButton = styled.button`
   width: 320px;
   height: 44px;
   border-radius: 5px;
@@ -36,21 +35,3 @@ const ConfirmButton = styled.button`
   font-size: ${({ theme }) => theme.fontSize.md};
   text-align: center;
 `;
-
-const CardRegistrationCompletePage = () => {
-  return (
-    <CardRegistrationCompleteLayout>
-      <CheckBox>
-        <img src={Check} />
-      </CheckBox>
-      <CardRegistrationMessage>
-        1234로 시작하는
-        <br />
-        BC카드가 등록되었어요.
-      </CardRegistrationMessage>
-      <ConfirmButton type="button">확인</ConfirmButton>
-    </CardRegistrationCompleteLayout>
-  );
-};
-
-export default CardRegistrationCompletePage;
